(refactor) Extract `BossLevel` from `levels`
